+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # NEXT RELEASE
+* UserData.followedTags is just an array of Tags
+* add some error handling in fetchList
 
 ### v0.21.4
 * Use the `/api/v1/timelines/tag/:hashtag` endpoint to pull trending tag toots in addition to search endpoint